fix: properly register view components, remove icons
