--- phpcs.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<ruleset name="PHP_CodeSniffer">
-    <description>PHPCS configuration file.</description>
-    <file>app</file>
-    <exclude-pattern>*/migrations/*</exclude-pattern>
-    <arg value="np"/>
-    <rule ref="PSR12"/>
-</ruleset>
