+++ resources/views/components/alert.blade.php
@@ -1,4 +1,4 @@
 <div class="alert alert-{{ $type }} flow" role="alert">
-    <p class="title">@svg($icon, ["aria-hidden" => "true"]) {{ $title }}</p>
+    <p class="title">{{ $title }}</p>
     {{ $slot }}
 </div>
