+++ src/HearthServiceProvider.php
@@ -3,6 +3,9 @@
 namespace Hearth;
 
 use Hearth\Commands\HearthCommand;
+use Hearth\View\Components\Alert;
+use Hearth\View\Components\LocaleSelect;
+use Hearth\View\Components\Select;
 use Illuminate\Support\Facades\Blade;
 use Illuminate\View\Compilers\BladeCompiler;
 use Spatie\LaravelPackageTools\Package;
@@ -39,10 +42,6 @@ class HearthServiceProvider extends PackageServiceProvider
             __DIR__ . '/../database/migrations/create_users_table.php.stub' =>
                 database_path('migrations/2014_10_12_000000_create_users_table.php'),
         ], 'hearth-migrations');
-
-        $this->publishes([
-            __DIR__ . '/Components' => app_path("View/Components/vendor/hearth"),
-        ], "hearth-components");
     }
 
     /**
@@ -60,6 +59,12 @@ class HearthServiceProvider extends PackageServiceProvider
             $this->registerComponent('locale-select');
             $this->registerComponent('select');
         });
+
+        $this->loadViewComponentsAs('hearth', [
+            Alert::class,
+            LocaleSelect::class,
+            Select::class,
+        ]);
     }
 
     /**
