+++ stubs/resources/views/components/validation-error.blade.php
@@ -1,3 +1,3 @@
 <p class="validation-error">
-    @svg("heroicon-s-exclamation", ["aria-hidden" => "true"]) {{ $slot }}
+    {{ $slot }}
 </p>
